feat,fix<BE> : 유료 API 조회 구현 및 헤더별 게시글 검색 자료형 수정

See merge request s07-webmobile1-sub2/S07P12B204!95

+++ backend/src/main/java/com/ssafy/cheertogether/article/service/ArticleService.java
@@ -60,7 +60,7 @@ public class ArticleService {
 	}
 
 	@Transactional(readOnly = true)
-	public List<ArticleResponse> findByHeader(int leagueApiId) {
+	public List<ArticleResponse> findByHeader(Long leagueApiId) {
 		return articleRepository
 			.findArticlesByLeague_ApiId(leagueApiId, Sort.by(Sort.Direction.DESC, "id"))
 			.stream()
